修改：计算转速时将时间单位设置为0.1ms，int16改为uint32，非低功耗下3个功能ok

+++ keil_projects/proj_template/proj_template/my_app/rev_app.h
@@ -8,7 +8,7 @@ typedef struct RevArg{
 	int32	revTimerCnt;	// Timer register CNT value
 	float	revTime;		// CNT value corresponding actualtime
 	int16	revFlag;		// execute TIMER_Reset or TIMER_Start
-	int16	targetRPM;		// target Revolutions Per minute
+	uint32	targetRPM;		// target Revolutions Per minute
 }RevArg_t;
